Split dataReducer in two
