+++ frontend/src/actions/index.js
@@ -1,5 +1,6 @@
 export * from './signInActions';
-export * from './dataActions';
+export * from './serviceProviderActions';
+export * from './censusDivisionActions';
 export * from './changeDropdownActions';
 export * from './changeSliderActions';
 export * from './changeViewActions';
