+++ frontend/src/actions/serviceProviderActions.js
@@ -1,6 +1,5 @@
 import {
   LOAD_SERVICE_PROVIDERS,
-  LOAD_CENSUS_DIVISION_DATA
 } from './actionsTypes';
 
 export const getServiceProviders = (token) => {
@@ -19,28 +18,6 @@ export const getServiceProviders = (token) => {
         dispatch({
           type: LOAD_SERVICE_PROVIDERS,
           service_providers: data
-        })
-      })
-      .catch((error) => {
-        console.log(error)
-      })
-  }
-}
-
-export const getCensusDivisionData = () => {
-  // TODO: add id_token authentication
-  return (dispatch) => {
-    fetch('/census_division_aggregate')
-      .then((response) => {
-        return response.json()
-      })
-      .then((json) => {
-        return json.data
-      })
-      .then((data) => {
-        dispatch({
-          type: LOAD_CENSUS_DIVISION_DATA,
-          censusDivisionData: data
         })
       })
       .catch((error) => {
