+++ frontend/src/components/Map.js
@@ -144,7 +144,7 @@ class Map extends Component {
 function mapStateToProps(state) {
   return {
     view: state.changeViewReducer.view,
-    serviceProviders: state.dataReducer.serviceProviders
+    serviceProviders: state.serviceProviderReducer
   };
 }
