+++ frontend/src/reducers/index.js
@@ -4,13 +4,15 @@ import changeDropdownReducer from './changeDropdownReducer';
 import changeSliderReducer from './changeSliderReducer';
 import selectCensusDivisionReducer from './selectCensusDivisionReducer'
 import authReducer from './authReducer';
-import dataReducer from './dataReducer';
+import serviceProviderReducer from './serviceProviderReducer';
+import censusDivisionReducer from './censusDivisionReducer';
 
 export default combineReducers({
-    changeViewReducer,
-    changeDropdownReducer,
-    changeSliderReducer,
-    authReducer,
-    dataReducer,
-    selectCensusDivisionReducer,
+  changeViewReducer,
+  changeDropdownReducer,
+  changeSliderReducer,
+  authReducer,
+  serviceProviderReducer,
+  censusDivisionReducer,
+  selectCensusDivisionReducer,
 });
